+++ GamePrograming/Application/Application.vcxproj
@@ -209,6 +209,7 @@
     <ClInclude Include="Source\Game\FieldManager.h" />
     <ClInclude Include="Source\Game\FieldObject.h" />
     <ClInclude Include="Source\Game\Kokeshi.h" />
+    <ClInclude Include="Source\Game\Phase.h" />
     <ClInclude Include="Source\Game\ThrowObject.h" />
     <ClInclude Include="Source\Game\GameObject.h" />
     <ClInclude Include="Source\Game\GameScene.h" />
@@ -234,6 +235,7 @@
     <ClCompile Include="Source\Game\FieldManager.cpp" />
     <ClCompile Include="Source\Game\FieldObject.cpp" />
     <ClCompile Include="Source\Game\Kokeshi.cpp" />
+    <ClCompile Include="Source\Game\Phaze.cpp" />
     <ClCompile Include="Source\Game\ThrowObject.cpp" />
     <ClCompile Include="Source\Game\GameObject.cpp" />
     <ClCompile Include="Source\Game\GameScene.cpp" />
