+++ GamePrograming/Application/Source/Game/GameScene.cpp
@@ -1,43 +1,28 @@
 /******************************************************
-* GameScene.cpp		ゲームシーン管理
+* GameScene.h		ゲームシーン管理
 * 制作者：ミヤタジョウジ
 * 作成日：2024/10/22
 * 最終更新日：2024/10/22
 *******************************************************/
 #include "framework.h"
 #include "DirectX/DirectX.h"
-#include "GameScene.h"
-#include "Game/Player.h"
+#include "Game/GameScene.h"
 
 /****************************************************
 * ゲームシーン初期化
 *****************************************************/
-GameScene::GameScene(const float gravityX, const float gravityY) : m_physics(new Physics(gravityX, gravityY)) {
-	m_state = GAMESCENESTATE_RUN;
-
+GameScene::GameScene() {
 	m_camera = new Camera();
-	m_fieldManager = new FieldManager();
-	m_throwObjectManager = new ThrowObjectManager();
-	m_player = new Player();
+	m_phaseNum = 0;
+	m_phase = new Phase(0.0f, 9.8f);
 }
 
 /****************************************************
 * ゲームシーン更新
 *****************************************************/
 void GameScene::Update() {
-	switch (m_state) {
-		case GAMESCENESTATE_AWAKE:
-			Awake();
-			break;
-		case GAMESCENESTATE_ASREEP:
-			Asreep();
-			break;
-		case GAMESCENESTATE_RUN:
-			Run();
-			break;
-		default:
-			break;
-	}
+	m_phase->Update();
+	m_camera->Update();
 }
 
 /****************************************************
@@ -45,44 +30,20 @@ void GameScene::Update() {
 *****************************************************/
 void GameScene::Draw() {
 	m_camera->Draw();
-
-	m_fieldManager->Draw();
-	m_throwObjectManager->Draw();
-	m_player->Draw();
+	m_phase->Draw();
 }
 
 /****************************************************
 * ゲームシーン終了
 *****************************************************/
 GameScene::~GameScene() {
-	if (m_physics) delete m_physics;
-	if (m_player) delete m_player;
-	if (m_fieldManager) delete m_fieldManager;
-	if (m_throwObjectManager) delete m_throwObjectManager;
+	if (m_phase) delete m_phase;
 	if (m_camera) delete m_camera;
 }
 
 /****************************************************
-* ゲームシーン起動
+* フェーズチェンジ
 *****************************************************/
-void GameScene::Awake() {
-	
-}
+void GameScene::ChangePhase() {
 
-/****************************************************
-* ゲームシーン終了まで
-*****************************************************/
-void GameScene::Asreep() {
-	
-}
-
-/****************************************************
-* ゲームシーン実行
-*****************************************************/
-void GameScene::Run() {
-	m_physics->UpdatePhysics((1.0f / 60.0f), 8, 3);
-	m_fieldManager->Update();
-	m_throwObjectManager->Update();
-	m_player->Update();
-	m_camera->Update();
 }
