+++ GamePrograming/Application/Source/Game/Phase.h
@@ -1,0 +1,48 @@
+/******************************************************
+* Phaseh		フェーズ管理
+* 制作者：ミヤタジョウジ
+* 作成日：2024/10/22
+* 最終更新日：2024/10/22
+*******************************************************/
+#pragma once
+
+#include "Physics.h"
+#include "Game/GameObject.h"
+#include "Game/FieldManager.h"
+#include "Game/ThrowObjectManager.h"
+
+enum GAMESCENESTATE {
+	GAMESCENESTATE_AWAKE = 0,
+	GAMESCENESTATE_ASREEP,
+	GAMESCENESTATE_RUN,
+};
+
+/****************************************************
+* フェーズクラス
+*****************************************************/
+class Phase {
+public:
+	Phase() = delete;
+	Phase(const float gravityX, const float gravityY);
+	~Phase();
+
+	virtual void Update();
+	virtual void Draw();
+
+	virtual void Awake();
+	virtual void Asreep();
+	virtual void Run();
+protected:
+	//世界のルール
+	Physics* m_physics = nullptr;
+
+	//フェーズのステート
+	GAMESCENESTATE m_state;
+
+	//マネージャー群
+	GameObject* m_player = nullptr;
+	FieldManager* m_fieldManager = nullptr;
+	ThrowObjectManager* m_throwObjectManager = nullptr;
+private:
+};
+
