+++ GamePrograming/Application/x64/Debug/Application.tlog/Cl.items.tlog
@@ -10,6 +10,7 @@ C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Ga
 C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\FieldManager.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\FieldManager.obj
 C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\FieldObject.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\FieldObject.obj
 C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\Kokeshi.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Kokeshi.obj
+C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\Phaze.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Phaze.obj
 C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\ThrowObject.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\ThrowObject.obj
 C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\GameObject.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\GameObject.obj
 C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\GameScene.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\GameScene.obj
