+++ GamePrograming/Application/x64/Debug/Application.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\APPLICATION.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\APPLICATION.RES|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\AUDIO.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\BEAR.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\CAMERA.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\CONTROLLER.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\DIRECT3D.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\FIELD.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\FIELDMANAGER.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\FIELDOBJECT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMEOBJECT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMESCENE.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMESYSTEM.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GROUND.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\INPUT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\KOKESHI.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\PHYSICS.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\PLAYER.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\TEXTURE.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\THROWOBJECT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\THROWOBJECTMANAGER.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\WOODENBOX.OBJ
+^C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\APPLICATION.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\APPLICATION.RES|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\AUDIO.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\BEAR.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\CAMERA.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\CONTROLLER.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\DIRECT3D.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\FIELD.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\FIELDMANAGER.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\FIELDOBJECT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMEOBJECT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMESCENE.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMESYSTEM.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GROUND.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\INPUT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\KOKESHI.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\PHAZE.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\PHYSICS.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\PLAYER.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\TEXTURE.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\THROWOBJECT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\THROWOBJECTMANAGER.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\WOODENBOX.OBJ
 C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Application.ilk
